Added an Example for Contact Format as JSON. I hope that this can be useful to my co-worker "Basso Enrico" to create a graphical form for contact info insertion too.
Please leave this file as you see it from this commit. If you need to modify it just do a copy and do it there.

I want to also point out that this is only a structure example when it will be implemented it could be used something different from JSON format.